changed from floating point arthm to integer

+++ nnet/nnet.h
@@ -16,24 +16,24 @@ namespace ML {
   //in_nodes := amount of input nodes
   //nodes    := amount of nodes in layer
   struct Layer {
-    Mat<f64> w;
-    Vec<f64> b;
-    Vec<f64> a;
+    Mat<u64> w;
+    Vec<u64> b;
+    Vec<u64> a;
 
     Layer(uint in_nodes, uint nodes);
-    Layer(uint in_nodes, uint nodes, f64 *buffer);
-    Vec<f64> *get_active(Vec<f64> &input);
+    Layer(uint in_nodes, uint nodes, u64 *buffer);
+    Vec<u64> *get_active(Vec<u64> &input);
   };
 
   struct Nnet_Structure : public vector<uint> {
     uint data_size();
-    vector<Layer> *build(f64 *data);
+    vector<Layer> *build(u64 *data);
     void load(FILE *file);
     void save(FILE *file);
   };
 
   struct Nnet {
-    f64            *data;
+    u64            *data;
     uint           data_size;
     Nnet_Structure structure;
     vector<Layer>  *lv;
@@ -42,13 +42,13 @@ namespace ML {
     Nnet(string &path);
     ~Nnet();
 
-    Vec<f64> *get_output(Vec<f64> &in);
+    Vec<u64> *get_output(Vec<u64> &in);
 
     void load(FILE *file);
     void save(FILE *file);
   };
 
-  f64 cost(Vec<f64> &v, Vec<f64> &exp);
+  u64 cost(Vec<u64> &v, Vec<u64> &exp);
 
   std::ostream &operator<<(std::ostream &os, Layer &l);
   std::ostream &operator<<(std::ostream& os, Nnet &n);
